+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Added
+
+- Added automatic apostrophe insertion in Katakana to Latin conversion.
+- Added -y (ey, oy, uy) support for diphthongs in Katakana to Latin conversion.
+
 ## [0.4.0] - 2024-03-09
 
 ### Added